+++ kvdb-web/CHANGELOG.md
@@ -24,4 +24,4 @@ The format is based on [Keep a Changelog].
 
 ## [0.1.1] - 2019-10-24
 ### Dependencies
-- Updated dependencies (https://github.com/paritytech/parity-common/pull/239)
+- Updated dependencies. [#239](https://github.com/paritytech/parity-common/pull/239)